tutorials: rework "input_from_gui" to split code into three different files.
